水平越权攻击修复

+++ xdclass-user-service/src/main/java/net/xdclass/service/impl/AddressServiceImpl.java
@@ -43,7 +43,10 @@ public class AddressServiceImpl implements AddressService {
      **/
     @Override
     public AddressVO detail(Long id) {
-        AddressDO addressDO = addressMapper.selectOne(new QueryWrapper<AddressDO>().eq("id", id));
+        LoginUser loginUser= LoginInterceptor.threadLocal.get();
+        AddressDO addressDO = addressMapper.selectOne(new QueryWrapper<AddressDO>()
+                .eq("id", id)
+                .eq("user_id",loginUser.getId()));
         if(addressDO==null){
             return null;
         }
@@ -94,7 +97,8 @@ public class AddressServiceImpl implements AddressService {
      **/
     @Override
     public int delete(int addressId) {
-        int rows=addressMapper.delete(new QueryWrapper<AddressDO>().eq("id",addressId));
+        LoginUser loginUser = LoginInterceptor.threadLocal.get();
+        int rows=addressMapper.delete(new QueryWrapper<AddressDO>().eq("id",addressId).eq("user_id",loginUser.getId()));
         return rows;
     }
     /**
